+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Fixed
+- Error for empty `restClientConfig` while using HTTP retries.
 
 ## [5.5.7] - 2025-12-09
 ### Changed